Added minio docker compose to test aws s3 module
