+++ .gitignore
@@ -14,3 +14,4 @@ out
 .idea_modules
 .bsp
 .env
+.minio.sys
