+++ modules/aws/src/test/scala/etlflow/AwsTestHelper.scala
@@ -1,0 +1,10 @@
+package etlflow
+
+import software.amazon.awssdk.regions.Region
+
+trait AwsTestHelper {
+  lazy val s3_bucket: String = "test"
+  lazy val s3_region: Region = Region.AP_SOUTH_1
+  val canonical_path: String = new java.io.File(".").getCanonicalPath
+  val file                   = s"$canonical_path/modules/aws/src/test/resources/input/ratings.parquet"
+}
